chore: adds Vulnerabilities badge
I recently ran an audit on this software so that we were able to include it into our project. We will continue to monitor the library and it's dependencies. We'd like to help show your users you care about security and are using Snyk to protect yourselves and your users by adding a badge to the project readme showing you don't have known vulnerabilities in your project.